Rename config::Pedal -> Effect.
