+++ src/config/mod.rs
@@ -1,11 +1,13 @@
-use crate::{pedal, Pedal as PedalTrait, Result};
+mod effect;
+
+use crate::{pedal, Result};
 use cpal::StreamConfig;
+use effect::Effect;
 use serde::Deserialize;
-use Pedal::*;
 
 #[derive(Debug, Deserialize)]
 pub struct Config {
-    pedals: Vec<Pedal>,
+    pedals: Vec<Effect>,
 }
 
 impl Config {
@@ -15,7 +17,7 @@ impl Config {
 
     pub fn default() -> Self {
         Self {
-            pedals: vec![Transparent],
+            pedals: vec![Effect::Transparent],
         }
     }
 
@@ -24,46 +26,5 @@ impl Config {
             .iter()
             .map(|pedal| pedal.to_pedal(stream_config))
             .collect()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(tag = "type")]
-enum Pedal {
-    Transparent,
-    Delay(DelayConfig),
-}
-
-impl Pedal {
-    fn to_pedal(&self, stream_config: &StreamConfig) -> Result<pedal::Boxed> {
-        Ok(match self {
-            Transparent => pedal::Transparent::new().boxed(),
-            Delay(delay_config) => delay_config.to_pedal(stream_config)?,
-        })
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub struct DelayConfig {
-    pub level: f32,
-    pub delay_ms: u32,
-    pub num: u32,
-}
-
-impl DelayConfig {
-    fn to_pedal(&self, stream_config: &StreamConfig) -> Result<pedal::Boxed> {
-        let mut pedals = (1..=self.num)
-            .map(|n| {
-                Ok(
-                    pedal::Delay::new(stream_config, self.level.powi(n as i32), self.delay_ms * n)?
-                        .boxed(),
-                )
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        let transparent = pedal::Transparent::new().boxed();
-        pedals.insert(0, transparent);
-
-        Ok(pedal::Split::new(pedals)?.boxed())
     }
 }
